requirements.txt
requests
pandas
SQLAlchemy
PyYAML
psycopg2-binary
streamlit
matplotlib
plotly
